regroup functions from DC analysis, and write tests for injections (compensated or not)
