+++ src/main/java/com/powsybl/openloadflow/sensi/AcSensitivityAnalysis.java
@@ -7,22 +7,22 @@
 package com.powsybl.openloadflow.sensi;
 
 import com.powsybl.commons.PowsyblException;
-import com.powsybl.iidm.network.Bus;
-import com.powsybl.iidm.network.Injection;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.math.matrix.DenseMatrix;
 import com.powsybl.math.matrix.MatrixFactory;
 import com.powsybl.openloadflow.OpenLoadFlowParameters;
 import com.powsybl.openloadflow.ac.equations.AcEquationSystem;
+import com.powsybl.openloadflow.ac.equations.AcEquationSystemCreationParameters;
 import com.powsybl.openloadflow.ac.equations.ClosedBranchSide1ActiveFlowEquationTerm;
 import com.powsybl.openloadflow.equations.*;
 import com.powsybl.openloadflow.network.LfBranch;
-import com.powsybl.openloadflow.network.LfBus;
 import com.powsybl.openloadflow.network.LfNetwork;
+import com.powsybl.openloadflow.network.PerUnit;
+import com.powsybl.openloadflow.network.util.ParticipatingElement;
+import com.powsybl.openloadflow.util.PropagatedContingency;
 import com.powsybl.sensitivity.SensitivityFactor;
 import com.powsybl.sensitivity.SensitivityValue;
-import com.powsybl.sensitivity.factors.BranchFlowPerInjectionIncrease;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.*;
@@ -45,88 +45,31 @@ public class AcSensitivityAnalysis extends AbstractSensitivityAnalysis {
         return branch;
     }
 
-    private List<SensitivityFactor> validateFactors(Network network, List<SensitivityFactor> factors, LfNetwork lfNetwork) {
-        return factors.stream().filter(factor -> {
-            if (factor instanceof BranchFlowPerInjectionIncrease) {
-                BranchFlowPerInjectionIncrease injectionIncrease = (BranchFlowPerInjectionIncrease) factor;
-
-                String branchId = injectionIncrease.getFunction().getBranchId();
-                String injectionId = injectionIncrease.getVariable().getInjectionId();
-
-                LfBranch branch = getLfBranch(lfNetwork, branchId);
-
-                // skip open branches
-                LfBus bus1 = branch.getBus1();
-                LfBus bus2 = branch.getBus2();
-                if (bus1 == null || bus2 == null) {
-                    return false;
-                }
-
-                Injection<?> injection = getInjection(network, injectionId);
-                Bus bus = injection.getTerminal().getBusView().getBus();
-
-                // skip disconnected injections
-                if (bus == null) {
-                    return false;
-                }
-
-                LfBus lfBus = lfNetwork.getBusById(bus.getId());
-                if (lfBus.isSlack()) {
-                    throw new PowsyblException("Cannot analyse sensitivity of slack bus");
-                }
-            } else {
-                throw new UnsupportedOperationException("Factor type '" + factor.getClass().getSimpleName() + "' not yet supported");
-            }
-            return true;
-        })
-        .collect(Collectors.toList());
-    }
-
-    private DenseMatrix initRhs(List<SensitivityFactor> factors, Network network, LfNetwork lfNetwork, EquationSystem equationSystem) {
-        DenseMatrix rhs = new DenseMatrix(equationSystem.getSortedEquationsToSolve().size(), factors.size());
-
-        for (int index = 0; index < factors.size(); index++) {
-            SensitivityFactor<?, ?> factor = factors.get(index);
-
-            if (factor instanceof BranchFlowPerInjectionIncrease) {
-                BranchFlowPerInjectionIncrease injectionIncrease = (BranchFlowPerInjectionIncrease) factor;
-                LfBus lfBus = getInjectionLfBus(network, lfNetwork, injectionIncrease);
-                if (lfBus.isSlack()) {
-                    // cannot inject on slack atm (will change with compensation)
-                    // this is because slack is Theta/V, so we do not have any P equation
-                    continue;
-                }
-                Equation pEquation = equationSystem.getEquation(lfBus.getNum(), EquationType.BUS_P).orElseThrow();
-
-                rhs.set(pEquation.getColumn(), index, 1);
+    protected void setBaseCaseSensitivityValues(List<SensitivityFactorGroup> factorGroups, DenseMatrix factorsState) {
+        for (SensitivityFactorGroup factorGroup : factorGroups) {
+            for (LfSensitivityFactor<ClosedBranchSide1ActiveFlowEquationTerm> factor : factorGroup.getFactors()) {
+                factor.setBaseCaseSensitivityValue(factor.getEquationTerm().calculate(factorsState, factorGroup.getIndex()));
             }
         }
-
-        return rhs;
     }
 
-    private List<SensitivityValue> calculateSensitivityValues(Network network, List<SensitivityFactor> factors,
+    private List<SensitivityValue> calculateSensitivityValues(Network network, List<SensitivityFactorGroup> factorGroups,
                                                               LfNetwork lfNetwork, EquationSystem equationSystem,
                                                               DenseMatrix states) {
-        List<SensitivityValue> sensitivityValues = new ArrayList<>(factors.size());
+        List<SensitivityValue> sensitivityValues = new ArrayList<>(factorGroups.size());
 
-        for (int index = 0; index < factors.size(); index++) {
-            SensitivityFactor<?, ?> factor = factors.get(index);
-            if (factor instanceof BranchFlowPerInjectionIncrease) {
-                BranchFlowPerInjectionIncrease injectionIncrease = (BranchFlowPerInjectionIncrease) factor;
-                LfBranch branch = lfNetwork.getBranchById(injectionIncrease.getFunction().getBranchId());
-                ClosedBranchSide1ActiveFlowEquationTerm eq = equationSystem.getEquationTerm(SubjectType.BRANCH, branch.getNum(), ClosedBranchSide1ActiveFlowEquationTerm.class);
-                sensitivityValues.add(new SensitivityValue(factor, eq.calculate(states, index), Double.NaN, Double.NaN));
+        for (SensitivityFactorGroup factorGroup : factorGroups) {
+            for (LfSensitivityFactor<?> factor : factorGroup.getFactors()) {
+                sensitivityValues.add(new SensitivityValue(factor.getFactor(), factor.getBaseSensitivityValue() * PerUnit.SB, Double.NaN, Double.NaN));
             }
         }
-
         return sensitivityValues;
     }
 
     /**
-     * https://people.montefiore.uliege.be/vct/elec0029/lf.pdf
+     * https://people.montefiore.uliege.be/vct/elec0029/lf.pdf / Equation 32 is transposed
      */
-    public Pair<List<SensitivityValue>, Map<String, List<SensitivityValue>>> analyse(Network network, List<SensitivityFactor> factors, LoadFlowParameters lfParameters,
+    public Pair<List<SensitivityValue>, Map<String, List<SensitivityValue>>> analyse(Network network, List<SensitivityFactor> factors, List<PropagatedContingency> contingencies, LoadFlowParameters lfParameters,
                                           OpenLoadFlowParameters lfParametersExt) {
         Objects.requireNonNull(network);
         Objects.requireNonNull(factors);
@@ -136,12 +79,35 @@ public class AcSensitivityAnalysis extends AbstractSensitivityAnalysis {
         // create LF network (we only manage main connected component)
         List<LfNetwork> lfNetworks = LfNetwork.load(network, lfParametersExt.getSlackBusSelector());
         LfNetwork lfNetwork = lfNetworks.get(0);
-
-        // filter invalid factors
-        List<SensitivityFactor> validFactors = validateFactors(network, factors, lfNetwork);
-
+        checkContingencies(lfNetwork, contingencies);
+        checkSensitivities(network, lfNetwork, factors);
+        checkLoadFlowParameters(lfParameters);
         // create AC equation system
-        EquationSystem equationSystem = AcEquationSystem.create(lfNetwork, new VariableSet());
+        EquationSystem equationSystem = AcEquationSystem.create(lfNetwork, new VariableSet(), new AcEquationSystemCreationParameters(false, true, false));
+
+        List<LfSensitivityFactor<ClosedBranchSide1ActiveFlowEquationTerm>> lfFactors = factors.stream().map(factor -> LfSensitivityFactor.create(factor, network, lfNetwork, equationSystem, ClosedBranchSide1ActiveFlowEquationTerm.class)).collect(Collectors.toList());
+
+        // index factors by variable group to compute a minimal number of states
+        List<SensitivityFactorGroup> factorGroups = createFactorGroups(network, lfFactors);
+        if (factorGroups.isEmpty()) {
+            return Pair.of(Collections.emptyList(), Collections.emptyMap());
+        }
+
+        // compute the participation for each injection factor (+1 on the injection and then -participation factor on all
+        // buses that contain elements participating to slack distribution
+        List<ParticipatingElement> participatingElements = null;
+        Map<String, Double> slackParticipationByBus;
+        if (lfParameters.isDistributedSlack()) {
+            participatingElements = getParticipatingElements(lfNetwork, lfParameters);
+            slackParticipationByBus = participatingElements.stream().collect(Collectors.toMap(
+                element -> element.getLfBus().getId(),
+                element -> -element.getFactor(),
+                Double::sum
+            ));
+        } else {
+            slackParticipationByBus = Collections.singletonMap(lfNetwork.getSlackBus().getId(), -1d);
+        }
+        computeInjectionFactors(slackParticipationByBus, factorGroups);
 
         // create jacobian matrix from current network state
         VoltageInitializer voltageInitializer = new PreviousValueVoltageInitializer();
@@ -151,13 +117,15 @@ public class AcSensitivityAnalysis extends AbstractSensitivityAnalysis {
         // ((DenseMatrix) j.getMatrix()).toJamaMatrix().inverse().transpose().print(1, 6);
 
         // otherwise, defining the rhs matrix will result in integer overflow
-        assert Integer.MAX_VALUE / (equationSystem.getSortedEquationsToSolve().size() * Double.BYTES) > validFactors.size();
+        assert Integer.MAX_VALUE / (equationSystem.getSortedEquationsToSolve().size() * Double.BYTES) > factorGroups.size();
         // initialize right hand side from valid factors
-        DenseMatrix rhs = initRhs(validFactors, network, lfNetwork, equationSystem);
+        DenseMatrix factorsStates = initFactorsRhs(lfNetwork, equationSystem, factorGroups); // this is the rhs for the moment
 
         // solve system
-        DenseMatrix states = solveTransposed(rhs, j);
+        solveTransposed(factorsStates, j);
+
+        setBaseCaseSensitivityValues(factorGroups, factorsStates);
         // calculate sensitivity values
-        return Pair.of(calculateSensitivityValues(network, factors, lfNetwork, equationSystem, states), Collections.emptyMap());
+        return Pair.of(calculateSensitivityValues(network, factorGroups, lfNetwork, equationSystem, factorsStates), Collections.emptyMap());
     }
 }
