+++ src/test/java/com/powsybl/openloadflow/network/FourBusNetworkFactory.java
@@ -84,6 +84,13 @@ public class FourBusNetworkFactory extends AbstractLoadFlowNetworkFactory {
         return network;
     }
 
+    public static Network createWithTransfoCompensed() {
+        Network network = createWithTransfo();
+        Bus b2 = network.getBusBreakerView().getBus("b2");
+        createGenerator(b2, "g2", 2);
+        return network;
+    }
+
     public static Network createWithTwoGeneratorsAtBus2() {
         Network network = create();
         Bus b2 = network.getBusBreakerView().getBus("b2");
